+++ bahamas/src/java/bahamas/services/DeleteDonation.java
@@ -86,7 +86,7 @@ public class DeleteDonation extends HttpServlet {
 
                 } else {
                     //Verified token
-                    int contactId = Validator.isIntValid(jobject.get("id").getAsString());
+                    int contactId = Validator.isIntValid(jobject.get("contact_id").getAsString());
                     ContactDAO cDAO = new ContactDAO();
 
                     Contact c = cDAO.retrieveContactById(contactId);
@@ -96,8 +96,15 @@ public class DeleteDonation extends HttpServlet {
                         out.println(gson.toJson(json));
                         return;
                     } else {
-                        int donation_id = Validator.isIntValid(jobject.get("donation_id").getAsString());
                         
+                        if (!cDAO.retrieveContactByUsername(username).isIsAdmin()) {
+                            json.addProperty("message", "fail");
+                            out.println(gson.toJson(json));
+                            return;
+                        }
+                        
+                        int donation_id = Validator.isIntValid(jobject.get("donation_id").getAsString());
+
                         if (DonationDAO.deleteDonation(donation_id)) {
                             json.addProperty("message", "success");
                             out.println(gson.toJson(json));
